rename package to be compliant with poetry install
