+++ tests/test_init.py
@@ -1,13 +1,13 @@
 import unittest
 
-import mailgun
-from mailgun.api import MailgunApi
+import mailgunsdk
+from mailgunsdk.api import MailgunApi
 
 
 class MailgunInitTestCase(unittest.TestCase):
     def test_api(self):
-        self.assertIsInstance(mailgun.api, MailgunApi)
+        self.assertIsInstance(mailgunsdk.api, MailgunApi)
 
     def test_initialize(self):
-        mailgun.initialize("api-key-xxx")
-        self.assertEqual(mailgun.api.api_key, "api-key-xxx")
+        mailgunsdk.initialize("api-key-xxx")
+        self.assertEqual(mailgunsdk.api.api_key, "api-key-xxx")
