+++ tests/test_ips.py
@@ -3,8 +3,8 @@ import unittest
 
 from mock import patch
 
-from mailgun.api import MailgunApi
-from mailgun.ips import IPs
+from mailgunsdk.api import MailgunApi
+from mailgunsdk.ips import IPs
 
 
 class IPsTestCase(unittest.TestCase):
